MOTECH-2639: Added input filter directive (#33)
* MOTEECH-2639: Start of the filter input directive

* MOTECH-2639: Working directive for set of checkboxes, without one wrapping a set

* MOTECH-2639: Got input filter group working as it should

* MOTECH-2639: Added input filter directive

Input filter group directive allows to filter a list with a set of checkboxes. It takes ng-model and a list of checkboxes which are controlled as an input group and wraps an input filter directive which holds a single checkbox behaviour.

* MOTECH-2639: Improved documentation

Minor changes in documentation for input filter group controller.